Back (#48)
* from my brain

* Try

* Updated requirements

* remove import

* Update

* Update travis

* Temporary Test

* view

* Testing

* SHA256

* kako login

* strange change on model

* Pariticipant user's profile id and name

* Join, accept meetings

* Accept/Reject meetings

* Split Views

* Split Views

* Delete views
