+++ backend/server/models.py
@@ -1,7 +1,6 @@
 from django.db import models
 from django.contrib.auth.models import User
 import re
-from django.db.models.functions import Sqrt
 from math import sqrt
 
 # Path to default image
@@ -107,15 +106,23 @@ class Comment(models.Model):
     def __str__(self):
         return self.comment_text
 
+    # For notification 1 : New comment for host
+    def save(self, *args, **kwargs):
+        notification = Notification(meeting=self.parent_meeting, profile=self.parent_meeting.host, notification = Notification.NOTIFICATION_NEW_COMMENT_FOR_HOST)
+        notification.save()
+        super().save(*args, **kwargs)
+
 # we should add url field.
 class Notification(models.Model):
     NOTIFICATION_NEW_APPLY = 0
     NOTIFICATION_NEW_COMMENT_FOR_HOST = 1
-    NOTIFICATION_CHOICES = [(NOTIFICATION_NEW_APPLY, 'new apply'), (NOTIFICATION_NEW_COMMENT_FOR_HOST, 'new comment for host')]
+    NOTIFICATION_APPLY_REJECTED = 2
+    NOTIFICATION_APPLY_APPROVED = 3
+    NOTIFICATION_CHOICES = [(NOTIFICATION_NEW_APPLY, 'new apply'), (NOTIFICATION_NEW_COMMENT_FOR_HOST, 'new comment for host'),(NOTIFICATION_APPLY_REJECTED, 'apply is rejected'),(NOTIFICATION_APPLY_APPROVED, 'apply is approved')]
 
-    profile = models.ForeignKey(Profile, on_delete=models.CASCADE)
+    profile = models.ForeignKey(Profile,on_delete=models.CASCADE)
     checked = models.BooleanField(default=False)
-    url = models.URLField(blank = True)
+    meeting = models.ForeignKey(Meeting, on_delete=models.CASCADE, null=True)
     notification = models.IntegerField(choices=NOTIFICATION_CHOICES)
 
     def __str__(self):
@@ -131,19 +138,29 @@ class Membership(models.Model):
     meeting = models.ForeignKey(Meeting, on_delete=models.CASCADE)
     created_at = models.DateTimeField(auto_now_add=True)
     status = models.IntegerField(choices=STATUS_CHOICES)
-    message = models.CharField(max_length = 500)
+    message = models.CharField(max_length = 500, null=True, blank=True)
 
     def __str__(self):
-        return self.message
+        return str(self.meeting.id) + '@' + str(self.profile.id)
 
     class Meta:
         unique_together = (
             ('profile', 'meeting')
         )
 
-    # For notification 1 : New apply
-    # We should add url.
+    # For notification 0 : New apply
+    # For notification 2 : Apply rejected
+    # For notification 3 : Apply approved
     def save(self, *args, **kwargs):
-        notification = Notification(profile=self.meeting.host, notification = Notification.NOTIFICATION_NEW_APPLY)
-        notification.save()
+        if(self.pk==None):
+            notification = Notification(meeting=self.meeting, profile=self.meeting.host, notification = Notification.NOTIFICATION_NEW_APPLY)
+            notification.save()
+        else:
+            if(self.status == self.STATUS_CHOICES[1][0]):
+                notification = Notification(meeting=self.meeting, profile=self.profile, notification = Notification.NOTIFICATION_APPLY_APPROVED)
+                notification.save()
+                print("Notify")
+            elif(self.status == self.STATUS_CHOICES[2][0]):
+                notification = Notification(meeting = self.meeting, profile = self.profile, notification = Notification.NOTIFICATION_APPLY_REJECTED)
+                notification.save()
         super().save(*args, **kwargs)
