+++ backend/server/serializers.py
@@ -8,7 +8,9 @@ class ProfileSerializer(serializers.ModelSerializer):
 
 class MeetingSerializer(serializers.ModelSerializer):
     nickname = serializers.ReadOnlyField(source='host.nickname')
-    
+    # participant_waiting = serializers.ReadOnlyField(source='participant.through_fields.filter(status=0)')
+    # participant_approved = serializers.ReadOnlyField(source='participant.through_fields.filter(status=1)')
+
     class Meta:
         model = Meeting
         fields = ('id', 'name', 'host', 'nickname', 'date', 'posted_date', 'participant', 'max_participant', 'deadline', 'region', 'photo', 'content', 'tag_set', 'status', 'open_chat', 'comment_set', 'membership_set', 'latitude', 'longitude')
@@ -32,7 +34,7 @@ class NotificationSerializer(serializers.ModelSerializer):
 class MembershipSerializer(serializers.ModelSerializer):
     class Meta:
         model = Membership
-        fields = ('profile', 'meeting', 'created_at', 'status', 'message')
+        fields = ('id', 'profile', 'meeting', 'created_at', 'status', 'message')
 
 class UserSerializer(serializers.ModelSerializer):
     profile = serializers.PrimaryKeyRelatedField(many=True, queryset=Profile.objects.all())
